feat: implement modules solution
